Remove use of BDParcellaire
