+++ configs/optimize/default.yaml
@@ -4,7 +4,6 @@ todo: "prepare+optimize+evaluate+update"  # for test data use prepare+evaluate+u
 results_output_dir: "/path/to/folder/"  # will contain best optimization trial and (optionnaly) updated las
 best_trial_pickle_path: "/path/to/best_trial.pkl"  # To save optimized parameters from Optuna
 input_bd_topo_shp: ${oc.env:INPUT_BD_TOPO_SHP_PATH}
-input_bd_parcellaire_shp: ${oc.env:INPUT_BD_PARCELLAIRE_SHP_PATH}
 
 study:
   n_trials: 300
